+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## UNRELEASED
 
+- Reply listeners now inherit the Bus's `verifyDomain`/`targetFrames`
+
 ## 6.1.0
 
 - feat(deps): Add [conventional commits](https://www.conventionalcommits.org/en/v1.0.0/) to the repo